Swap experimental component for the real deal
