+++ src/components/Card/EmptyStateCard.js
@@ -1,151 +1,53 @@
 import React from 'react'
-import PropTypes from '../../proptypes'
-import styled from 'styled-components'
-import { theme } from '../../theme-legacy'
-import { noop, warn } from '../../utils'
+import PropTypes from 'prop-types'
 import { Button } from '../Button/Button'
-import Text from '../Text/Text'
 import Card from './Card'
-import omit from 'lodash-es/omit'
+import { GU, textStyle } from '../../style'
+import { useTheme } from '../../theme'
 
-class EmptyStateCard extends React.Component {
-  static propTypes = {
-    action: PropTypes.element,
-    actionButton: PropTypes._component,
-    actionDisabled: PropTypes.bool,
-    actionText: PropTypes.string,
-    icon: PropTypes.oneOfType([PropTypes.element, PropTypes._component]),
-    onActivate: PropTypes.func,
-    text: PropTypes.string,
-    title: PropTypes.string,
-  }
-  static defaultProps = {
-    title: 'Nothing here.',
-    actionDisabled: false,
-    onActivate: noop,
-  }
+const EmptyStateCard = React.memo(function EmptyStateCard({
+  actionButton,
+  actionText,
+  icon,
+  onClick,
+  text,
+}) {
+  const theme = useTheme()
 
-  constructor(props) {
-    super(props)
-    this.displayWarnings({}, props)
-  }
+  return (
+    <Card
+      css={`
+        display: grid;
+        grid-template-columns: 1fr;
+        grid-template-rows: auto 1fr auto;
+        padding: ${2.5 * GU}px ${2 * GU}px ${2 * GU}px;
+        text-align: center;
+      `}
+    >
+      {icon}
+      <span
+        css={`
+          color: ${theme.content};
+          ${textStyle('title4')};
+        `}
+      >
+        {text}
+      </span>
+      {actionButton && (
+        <Button onClick={onClick} wide mode="strong">
+          {actionText}
+        </Button>
+      )}
+    </Card>
+  )
+})
 
-  componentDidUpdate(prevProps) {
-    this.displayWarnings(prevProps, this.props)
-  }
-
-  displayWarnings(prevProps, props) {
-    // Deprecated warnings
-    if (prevProps.actionButton !== props.actionButton && props.actionButton) {
-      warn(
-        'The “actionButton” prop is deprecated. Please use “action” to pass ' +
-          'a custom element instead.'
-      )
-    }
-    if (
-      prevProps.icon !== props.icon &&
-      props.icon &&
-      !React.isValidElement(props.icon)
-    ) {
-      warn(
-        'Passing a component to the “icon” prop is deprecated. Please pass ' +
-          'an element instead (i.e. use icon={<Icon />} rather than ' +
-          'icon={Icon}).'
-      )
-    }
-
-    // Display a warning if `action` is used to override the button, while also
-    // using one of the props that apply to the default button component.
-    const anyActionParam =
-      props.actionDisabled || props.actionText || props.onActivate
-    if (prevProps.action !== props.action && props.action && anyActionParam) {
-      warn(
-        'Using `action` takes priority over `actionDisabled`, `actionText`, ' +
-          'and `onActivate`. Please use either `action` or any of these ' +
-          'props.'
-      )
-    }
-  }
-
-  // Render icon, even when provided as a component (deprecated)
-  iconElement() {
-    const { icon } = this.props
-    return !icon || React.isValidElement(icon)
-      ? icon
-      : React.createElement(icon)
-  }
-
-  // Render the action (usually a button)
-  actionElement() {
-    const {
-      action,
-      actionButton,
-      actionDisabled,
-      actionText,
-      onActivate,
-    } = this.props
-
-    // If an action node is provided, it takes priority above the other props
-    if (action) {
-      return action
-    }
-
-    // If no label is provided, the button is not displayed
-    if (!actionText) {
-      return null
-    }
-
-    // Use the provided actionButton if it exists, or a local component.
-    const Button = actionButton || DefaultActionButton
-
-    return (
-      <Button disabled={actionDisabled} mode="strong" onClick={onActivate} wide>
-        {actionText}
-      </Button>
-    )
-  }
-
-  render() {
-    const { title, text, ...props } = this.props
-    const extraProps = omit(props, Object.keys(EmptyStateCard.propTypes))
-    return (
-      <Main {...extraProps}>
-        <section>
-          {this.iconElement()}
-          <Heading>
-            <Text color={theme.accent} weight="bold" size="large">
-              {title}
-            </Text>
-          </Heading>
-          {text && <Text.Block>{text}</Text.Block>}
-          <ActionWrapper>{this.actionElement()}</ActionWrapper>
-        </section>
-      </Main>
-    )
-  }
+EmptyStateCard.propTypes = {
+  actionButton: PropTypes.bool,
+  actionText: PropTypes.string,
+  icon: PropTypes.node.isRequired,
+  onClick: PropTypes.func.isRequired,
+  text: PropTypes.node,
 }
-
-const Main = styled(Card)`
-  display: flex;
-  padding: 40px 60px;
-  align-items: center;
-  justify-content: center;
-  text-align: center;
-  section {
-    padding-top: 20px;
-  }
-`
-
-const DefaultActionButton = styled(Button)`
-  width: 150px;
-`
-
-const Heading = styled.h1`
-  margin: 20px 0 5px;
-`
-
-const ActionWrapper = styled.div`
-  margin-top: 20px;
-`
 
 export default EmptyStateCard
